pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = "Infrastructure management"
authors = ["Johnny Wezel <j@wezel.name>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
better-exceptions = "^0.3.3"
rich = "^13.8.0"
omegaconf = "^2.3.0"
fire = "^0.6.0"
plumbum = "^1.8.3"
dnspython = "^2.6.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
isort = "^5.13.2"
black = "^24.8.0"
flake8 = "^7.1.1"
snoop = "^0.4.3"
pdbpp = "^0.10.3"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
